refactor: continue components relocation for src/app
