+++ src/app/_components/main-wrapper.tsx
@@ -27,7 +27,7 @@ export default function MainContentWrapper({ children, title = null }: MainConte
             className={cn(
                 'flex flex-1 flex-col overflow-auto px-safe-or-6',
                 'mt-24 pb-safe', // Adjust top padding to account for topbar height
-                isBottomBarVisible ? 'mb-24' : '', // Add bottom padding when bottombar is visible
+                isBottomBarVisible ? 'mb-24' : '', // Add bottom margin when bottombar is visible
             )}>
             {children}
         </main>
