+++ src/app/_components/page-transitions/variants.ts
